model trained

+++ src/main/java/com/example/demo/NameMatchController.java
@@ -1,6 +1,5 @@
 package com.example.demo;
 
- 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -10,15 +9,18 @@ import java.util.Map;
 @RestController
 @RequestMapping("/api")
 public class NameMatchController {
-	@Autowired
-     private NameMatcherService nameMatcherService;
-     
+
+    @Autowired
+    private NameMatcherService nameMatcherService;
 
     @PostMapping("/match")
     public Map<String, Object> matchNames(@RequestBody Map<String, String> names) throws IOException {
         String correctName = names.get("correctName");
         String givenName = names.get("givenName");
-        int similarityScore = nameMatcherService.matchNames(correctName, givenName);
-        return Map.of("similarity", similarityScore + "%");
+
+        // Call service to compute similarity
+        double similarityScore = nameMatcherService.matchNames(correctName, givenName);
+
+        return Map.of("similarity", similarityScore +"%");
     }
 }
